requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
openai==1.13.3
httpx==0.27.2
python-dotenv==1.0.1